Updated section on comparison with Oph IRS 48, based on updated values I got from Nienke van der Marel (as opposed to the values we were using from Brown et al. 2012). The density is increased by a factor 10, and the particle radius decreased by a factor 3.
The turbulent velocities fall to 4% of the sound speed, SQUARELY what is predicted by the elliptic instability.

The trapped mass goes up to 6 Earth masses for Kida, and 17 for GNG vortices. The measured valued is 9! An EXCELLENT agreement, within a factor 2!

I don't have enough adjectives to say how nice this is!